Data analysis of fuel pricing
I completed this exercise for an interview. There were two components: a pricing task and a visualisation task, both with different data. The data provided was dummy data. The code is provided here, but not the data. I have provided links to the pricing analysis presentation and the visualisation spreadsheet.